+++ src/pyvism/backend/map.py
@@ -1,7 +1,7 @@
 """
 Contains a simple enum that lists the front-ends.
 """
-from pyvism.frontend import vmbc
+from pyvism.backend import vmbc
 from pyvism.py_utils import MapLikeEnum
