+++ src/pyvism/backend/vmbc/__init__.py
@@ -1,6 +1,6 @@
 """
 Virtual Machine ByteCode (VMBC) front-end for Vism.
 """
-from pyvism.frontend.vmbc.frontend import frontend
+from pyvism.backend.vmbc.frontend import frontend
 
 __all__ = ("frontend",)
