+++ src/pyvism/backend/vmbc/frontend.py
@@ -1,37 +1,37 @@
 from typing import LiteralString
 
-from pyvism.backend.instructions import ADD
-from pyvism.backend.instructions import DIFF
-from pyvism.backend.instructions import DIVMOD
-from pyvism.backend.instructions import INTDIV
-from pyvism.backend.instructions import MEMCH
-from pyvism.backend.instructions import MODULO
-from pyvism.backend.instructions import MUL
-from pyvism.backend.instructions import PATHJOIN
-from pyvism.backend.instructions import PRINTV
-from pyvism.backend.instructions import REPLIC
-from pyvism.backend.instructions import SFLUSH
-from pyvism.backend.instructions import SUB
-from pyvism.backend.instructions import SWRITE
-from pyvism.backend.instructions import UNION
-from pyvism.backend.tools import IRI
-from pyvism.frontend.tools import TargetFrontEnd
-from pyvism.frontend.vmbc.instructions import add
-from pyvism.frontend.vmbc.instructions import divmod
-from pyvism.frontend.vmbc.instructions import flush
-from pyvism.frontend.vmbc.instructions import intdiv
-from pyvism.frontend.vmbc.instructions import modulo
-from pyvism.frontend.vmbc.instructions import mov
-from pyvism.frontend.vmbc.instructions import mul
-from pyvism.frontend.vmbc.instructions import pathjoin
-from pyvism.frontend.vmbc.instructions import print
-from pyvism.frontend.vmbc.instructions import seqdiff
-from pyvism.frontend.vmbc.instructions import strdiff
-from pyvism.frontend.vmbc.instructions import sub
-from pyvism.frontend.vmbc.instructions import union
-from pyvism.frontend.vmbc.instructions import write
-from pyvism.frontend.vmbc.tools import AnyInstruction
-from pyvism.frontend.vmbc.tools import mnemonic
+from pyvism.backend.tools import TargetFrontEnd
+from pyvism.backend.vmbc.instructions import add
+from pyvism.backend.vmbc.instructions import divmod
+from pyvism.backend.vmbc.instructions import flush
+from pyvism.backend.vmbc.instructions import intdiv
+from pyvism.backend.vmbc.instructions import modulo
+from pyvism.backend.vmbc.instructions import mov
+from pyvism.backend.vmbc.instructions import mul
+from pyvism.backend.vmbc.instructions import pathjoin
+from pyvism.backend.vmbc.instructions import print
+from pyvism.backend.vmbc.instructions import seqdiff
+from pyvism.backend.vmbc.instructions import strdiff
+from pyvism.backend.vmbc.instructions import sub
+from pyvism.backend.vmbc.instructions import union
+from pyvism.backend.vmbc.instructions import write
+from pyvism.backend.vmbc.tools import AnyInstruction
+from pyvism.backend.vmbc.tools import mnemonic
+from pyvism.frontend.instructions import ADD
+from pyvism.frontend.instructions import DIFF
+from pyvism.frontend.instructions import DIVMOD
+from pyvism.frontend.instructions import INTDIV
+from pyvism.frontend.instructions import MEMCH
+from pyvism.frontend.instructions import MODULO
+from pyvism.frontend.instructions import MUL
+from pyvism.frontend.instructions import PATHJOIN
+from pyvism.frontend.instructions import PRINTV
+from pyvism.frontend.instructions import REPLIC
+from pyvism.frontend.instructions import SFLUSH
+from pyvism.frontend.instructions import SUB
+from pyvism.frontend.instructions import SWRITE
+from pyvism.frontend.instructions import UNION
+from pyvism.frontend.tools import IRI
 
 
 # PascalCase because FrontEnd is secretly a class
