+++ src/pyvism/backend/vmbc/tools.py
