+++ src/pyvism/vm/runner.py
@@ -1,8 +1,8 @@
 from typing import TextIO
 
+from pyvism.backend.map import CompilationTarget
 from pyvism.compiler.compiler import compile
 from pyvism.errsys.tools import report_abortion
-from pyvism.frontend.map import CompilationTarget
 from pyvism.vm.vm import VM
 from result import Err
 from result import Ok
